:recycle: Complete refactor
